requirements.txt
asgiref==3.4.1
autopep8==1.6.0
distlib==0.3.4
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
filelock==3.4.2
flake8==4.0.1
mccabe==0.6.1
numpy==1.21.4
pandas==1.3.4
platformdirs==2.4.1
psycopg2==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
virtualenv==20.13.0
